Refactor: 로그인, 회원가입 UserController 통합, UserReqDto 통합
