--- src/main/java/shop/mtcoding/job/controller/LoginController.java
@@ -1,59 +1,0 @@
-package shop.mtcoding.job.controller;
-
-import javax.servlet.http.HttpSession;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-
-import shop.mtcoding.job.dto.user.UserReqDto.LoginReqDto;
-import shop.mtcoding.job.handler.exception.CustomException;
-import shop.mtcoding.job.model.user.User;
-import shop.mtcoding.job.model.user.UserRepository;
-import shop.mtcoding.job.service.LoginService;
-
-@Controller
-public class LoginController {
-    @Autowired
-    private LoginService loginService;
-
-    @Autowired
-    private HttpSession session;
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @GetMapping("/loginForm")
-    public String loginForm() {
-        return "/login/loginForm";
-    }
-
-    @PostMapping("/login")
-    public String login(LoginReqDto loginReqDto) {
-        if (loginReqDto.getUsername() == null || loginReqDto.getUsername().isEmpty()) {
-            throw new CustomException("아이디를 작성해주세요");
-        }
-        if (loginReqDto.getPassword() == null || loginReqDto.getPassword().isEmpty()) {
-            throw new CustomException("비밀번호를 작성해주세요");
-        }
-        // 1. 로그인하기 service
-        User principal = loginService.로그인하기(loginReqDto);
-
-        // 2. session에 저장
-        session.setAttribute("principal", principal);
-
-        // 3. principal 유효성 검사
-        if (session.getAttribute("principal") == null) {
-            throw new CustomException("존재하지 않는 아이디거나 비밀번호를 다시 확인해주시기 바랍니다");
-        }
-
-        return "redirect:/";
-    }
-
-    @GetMapping("/logout")
-    public String logout() {
-        session.invalidate();
-        return "redirect:/";
-    }
-}
