+++ src/main/java/shop/mtcoding/job/dto/user/UserReqDto.java
@@ -6,7 +6,21 @@ import lombok.Setter;
 public class UserReqDto {
     @Setter
     @Getter
-    public static class JoinReqDto {
+    public static class LoginUserReqDto {
+        private String username;
+        private String password;
+    }
+
+    @Setter
+    @Getter
+    public static class LoginEnterpriseReqDto {
+        private String enterpriseName;
+        private String password;
+    }
+
+    @Setter
+    @Getter
+    public static class JoinUserReqDto {
         private String username;
         private String password;
         private String name;
@@ -17,8 +31,15 @@ public class UserReqDto {
 
     @Setter
     @Getter
-    public static class LoginReqDto {
-        private String username;
+    public static class JoinEnterpriseReqDto {
+        private String enterpriseName;
         private String password;
+        private String address;
+        private String contact;
+        private String image;
+        private String email;
+        private String sector;
+        private String size;
     }
+
 }
